git-svn-id: https://svn.kuali.org/repos/foundation/sandbox/maven-spring-mvc@29605 eb505253-de9c-4f3b-bf87-3aad84f29c0f

+++ src/main/java/springapp/service/SimpleProductManager.java
@@ -18,7 +18,7 @@ public class SimpleProductManager implements ProductManager {
 	public void increasePrice(int percentage) {
 		if (products != null) {
 			for (Product product : products) {
-				double newPrice = product.getPrice().doubleValue() * (100 + percentage) / 100;
+				double newPrice = product.getPrice() * (100 + percentage) / 100;
 				product.setPrice(newPrice);
 			}
 		}
